+++ src/Services/questionService.ts
@@ -1,14 +1,10 @@
 import { nanoid } from 'nanoid'
 import {
   MovieCast,
-  MovieCreditsResult,
   MovieCrew,
-  MovieDetailsResult,
   MovieProductionCompanies,
-  MovieResult,
   QuizQuestion,
 } from '../globalTypes'
-import { getMovieCredits, getMovieDetails } from './movieDatabaseService'
 import {
   getWrongActors,
   getWrongCharacters,
@@ -198,27 +194,11 @@ const askCharacterIsActor = (cast: MovieCast[]): QuizQuestion => {
   return newQuestion
 }
 
-const createNewQuiz = async (data: MovieResult): Promise<QuizQuestion[]> => {
-  const detailsData = (await getMovieDetails(
-    String(data.id),
-  )) as MovieDetailsResult
-
-  const creditsData = (await getMovieCredits(
-    String(data.id),
-  )) as MovieCreditsResult
-
-  if (creditsData && detailsData) {
-    const q1 = askReleaseYear(data.release_date)
-    const q2 = askDirector(creditsData.crew)
-    const q3 = askProductionCompany(detailsData.production_companies)
-    const q4 = askMovieBudget(detailsData.budget)
-    const q5 = askActorPlaysCharacter(creditsData.cast)
-    const q6 = askCharacterIsActor(creditsData.cast)
-
-    return [q1, q2, q3, q4, q5, q6]
-  } else {
-    return []
-  }
+export {
+  askReleaseYear,
+  askDirector,
+  askProductionCompany,
+  askMovieBudget,
+  askActorPlaysCharacter,
+  askCharacterIsActor,
 }
-
-export { createNewQuiz }
